+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 1.0.0
 
 - Initial version.
+- sounfont parser returns only information from header like version and list of presets and instruments